TexturedCubeJavaExample as a JUnit test
